factored out the table and graph rendering and applied it to datas view as well
