+++ public/views/Datas.js
@@ -1,14 +1,11 @@
-import Graph from "../components/graph/Graph.js";
-import Actions from "../store/Actions.js";
+import { getDataAsGraph, setupToolBar} from "../helpers/dataRendererHelper.js";
 
 export default class Datas {
   constructor() {
     document.title = "Data";
+    this.returnRenderFunc = getDataAsGraph;
+    this.view = "datas";
   }
-
-  async getTemplate() {
-    const view = "datas";
-    await Actions.GETALL(view);
-    return Graph(view);
-  }
+  async getTemplate() { return this.returnRenderFunc("datas") }
+  async setupToolBar() { return setupToolBar("datas") }
 }
